third_party/WebKit/LayoutTests/svg/animations/discard-on-discard-expected.txt

SVG 1.1 dynamic animation tests

Test the behavior of one discard applied on another discard

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS successfullyParsed is true

TEST COMPLETE
